dataengineering/challenge-3/streamer/requirements.txt
kafka-python==2.0.2
requests==2.24.0
requests_oauthlib==1.3.0